feat: implemented update logic for deliveries
