+++ frontend/src/contexts/dataModalContext.tsx
@@ -31,6 +31,7 @@ import { Demand } from '@models/types/data/demand';
 import { Production } from '@models/types/data/production';
 import { Material, Stock, StockType } from '@models/types/data/stock';
 import { DirectionType } from '@models/types/erp/directionType';
+import { DeliveryCreationModal } from '@features/material-details/components/DeliveryCreationModal';
 
 export type DataCategory = 'production' | 'demand' | 'stock' | 'delivery';
 
@@ -109,10 +110,11 @@ export const DataModalProvider = ({ children, material }: DataModalProviderProps
 
     const openDeliveryDialog = useCallback(
         (d: Partial<Delivery>, mode: ModalMode, direction: DirectionType, site: Site | null, list: Delivery[]) => {
+            const optionName = mode === 'view' ? 'viewDeliveryDialogOptions' : 'editDeliveryDialogOptions';
             d.ownMaterialNumber = materialNumber;
             dispatch({ type: 'delivery', payload: d });
             dispatch({ type: 'deliveries', payload: list });
-            dispatch({ type: 'deliveryDialogOptions', payload: { open: true, mode, direction, site } });
+            dispatch({ type: optionName, payload: { open: true, direction, site } });
         },
         [materialNumber]
     );
@@ -156,7 +158,7 @@ export const DataModalProvider = ({ children, material }: DataModalProviderProps
     );
     return (
         <>
-            <dataModalContext.Provider value={{ openDialog, addOnSaveListener, removeOnSaveListener }}>{children}</dataModalContext.Provider>
+            <dataModalContext.Provider value={{ openDialog, addOnSaveListener, removeOnSaveListener }}>{children}
             <DemandCategoryModal
                 {...state.demandDialogOptions}
                 onClose={() => dispatch({ type: 'demandDialogOptions', payload: { open: false, mode: state.demandDialogOptions.mode } })}
@@ -182,12 +184,24 @@ export const DataModalProvider = ({ children, material }: DataModalProviderProps
                 productions={state.productions}
             />
             <DeliveryInformationModal
-                {...state.deliveryDialogOptions}
-                onClose={() => dispatch({ type: 'deliveryDialogOptions', payload: { ...state.deliveryDialogOptions, open: false } })}
+                {...state.viewDeliveryDialogOptions}
+                onClose={() => dispatch({ type: 'viewDeliveryDialogOptions', payload: { ...state.viewDeliveryDialogOptions, open: false } })}
                 onSave={() => onSave('delivery')}
                 onRemove={(deletedUuid: string) => {
                     const updatedDeliveries = state.deliveries.filter(p => p.uuid !== deletedUuid);
                     dispatch({ type: 'deliveries', payload: updatedDeliveries });
+                    onSave('delivery');
+                }}
+                delivery={state.delivery}
+                deliveries={state.deliveries}
+            />
+            <DeliveryCreationModal
+                {...state.editDeliveryDialogOptions}
+                onClose={() => dispatch({ type: 'editDeliveryDialogOptions', payload: { ...state.editDeliveryDialogOptions, open: false } })}
+                onSave={(delivery) => {
+                    const updatedDeliveries = state.deliveries.map(d => d.uuid === delivery?.uuid ? delivery! : d);
+                    dispatch({ type: 'deliveries', payload: updatedDeliveries });
+                    dispatch({ type: 'delivery', payload: delivery ?? null });
                     onSave('delivery');
                 }}
                 delivery={state.delivery}
@@ -205,7 +219,7 @@ export const DataModalProvider = ({ children, material }: DataModalProviderProps
                 stock={state.stock}
                 stocks={state.stocks}
             />
-
+            </dataModalContext.Provider>
         </>
     );
 };
@@ -219,7 +233,8 @@ export function useDataModal() {
 }
 
 type ModalState = {
-    deliveryDialogOptions: { open: boolean; mode: ModalMode; direction: DirectionType; site: Site | null };
+    editDeliveryDialogOptions: { open: boolean; direction: DirectionType; site: Site | null };
+    viewDeliveryDialogOptions: { open: boolean; direction: DirectionType; site: Site | null };
     demandDialogOptions: { open: boolean; mode: ModalMode };
     productionDialogOptions: { open: boolean; mode: ModalMode };
     stockDialogOptions: { open: boolean; mode: ModalMode, stockType: StockType };
@@ -243,7 +258,8 @@ const reducer = (state: ModalState, action: ModalAction): ModalState => {
 };
 
 const initialState: ModalState = {
-    deliveryDialogOptions: { open: false, mode: 'edit', direction: DirectionType.Inbound, site: null },
+    editDeliveryDialogOptions: { open: false, direction: DirectionType.Inbound, site: null },
+    viewDeliveryDialogOptions: { open: false, direction: DirectionType.Inbound, site: null },
     demandDialogOptions: { open: false, mode: 'edit' },
     productionDialogOptions: { open: false, mode: 'edit' },
     stockDialogOptions: { open: false, mode: 'edit', stockType: 'material' },
