+++ frontend/src/features/material-details/components/DeliveryInformationModal.tsx
@@ -18,27 +18,20 @@ under the License.
 
 SPDX-License-Identifier: Apache-2.0
 */
-import { Input, Table } from '@catena-x/portal-shared-components';
-import { DateTime } from '@components/ui/DateTime';
-import { usePartners } from '@features/stock-view/hooks/usePartners';
-import { UNITS_OF_MEASUREMENT } from '@models/constants/uom';
+import { Table } from '@catena-x/portal-shared-components';
 import { Delivery } from '@models/types/data/delivery';
-import { Close, Delete, Save } from '@mui/icons-material';
-import { Box, Button, Dialog, DialogTitle, FormLabel, Grid, Stack, capitalize } from '@mui/material';
-import { deleteDelivery, postDelivery } from '@services/delivery-service';
-import { getUnitOfMeasurement, isValidOrderReference } from '@util/helpers';
+import { Close, Edit,  Delete } from '@mui/icons-material';
+import { Box, Button, Dialog, DialogTitle, Grid, Stack } from '@mui/material';
+import { deleteDelivery } from '@services/delivery-service';
+import { getUnitOfMeasurement } from '@util/helpers';
 import { useEffect, useMemo, useState } from 'react';
 import { INCOTERMS } from '@models/constants/incoterms';
-import { ARRIVAL_TYPES, DEPARTURE_TYPES } from '@models/constants/event-type';
-import { ModalMode } from '@models/types/data/modal-mode';
 import { Site } from '@models/types/edc/site';
-import { LabelledAutoComplete } from '@components/ui/LabelledAutoComplete';
-import { GridItem } from '@components/ui/GridItem';
-import { useSites } from '@features/stock-view/hooks/useSites';
 import { useNotifications } from '@contexts/notificationContext';
 import { DirectionType } from '@models/types/erp/directionType';
+import { useDataModal } from '@contexts/dataModalContext';
 
-const createDeliveryColumns = (handleDelete: (row: Delivery) => void) => {
+const createDeliveryColumns = (handleDelete: (row: Delivery) => void, handleEdit: (row: Delivery) => void) => {
     return [
         {
             field: 'dateOfDeparture',
@@ -190,7 +183,7 @@ const createDeliveryColumns = (handleDelete: (row: Delivery) => void) => {
             field: 'lastUpdatedOnDateTime',
             headerName: 'Updated',
             headerAlign: 'center',
-            flex: 1.5,
+            width: 100,
             renderCell: (data: { row: Delivery }) => (
                 <Stack display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
                     <Box>{new Date(data.row.lastUpdatedOnDateTime).toLocaleDateString('en-GB')}</Box>
@@ -199,13 +192,33 @@ const createDeliveryColumns = (handleDelete: (row: Delivery) => void) => {
             ),
         },
         {
+            field: 'edit',
+            headerName: '',
+            sortable: false,
+            disableColumnMenu: true,
+            headerAlign: 'center',
+            type: 'string',
+            width: 20,
+            renderCell: (data: { row: Delivery }) => {
+                return (
+                    <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                        {!data.row.reported && (
+                            <Button variant="text" onClick={() => handleEdit(data.row)} data-testid="edit-delivery">
+                                <Edit></Edit>
+                            </Button>
+                        )}
+                    </Box>
+                );
+            },
+        },
+        {
             field: 'delete',
             headerName: '',
             sortable: false,
             disableColumnMenu: true,
             headerAlign: 'center',
             type: 'string',
-            width: 30,
+            width: 20,
             renderCell: (data: { row: Delivery }) => {
                 return (
                     <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
@@ -221,31 +234,12 @@ const createDeliveryColumns = (handleDelete: (row: Delivery) => void) => {
     ] as const;
 };
 
-const isValidDelivery = (delivery: Partial<Delivery>) =>
-    delivery &&
-    delivery.ownMaterialNumber &&
-    delivery.originBpns &&
-    delivery.partnerBpnl &&
-    delivery.destinationBpns &&
-    typeof delivery.quantity === 'number' && delivery.quantity > 0 &&
-    delivery.measurementUnit &&
-    delivery.incoterm &&
-    delivery.dateOfDeparture &&
-    delivery.dateOfArrival &&
-    delivery.departureType &&
-    delivery.arrivalType &&
-    delivery.dateOfArrival >= delivery.dateOfDeparture &&
-    (delivery.departureType !== 'actual-departure' || delivery.dateOfDeparture <= new Date()) &&
-    (delivery.arrivalType !== 'actual-arrival' || delivery.dateOfArrival <= new Date()) &&
-    isValidOrderReference(delivery);
-
-type DeliveryInformationModalProps = {
+export type DeliveryInformationModalProps = {
     open: boolean;
-    mode: ModalMode;
     direction: DirectionType;
     site: Site | null;
     onClose: () => void;
-    onSave: () => void;
+    onSave: (d?: Delivery) => void;
     onRemove?: (deletedUuid: string) => void;
     delivery: Delivery | null;
     deliveries: Delivery[];
@@ -253,20 +247,16 @@ type DeliveryInformationModalProps = {
 
 export const DeliveryInformationModal = ({
     open,
-    mode,
     direction,
     site,
     onClose,
-    onSave,
     onRemove,
     delivery,
     deliveries,
 }: DeliveryInformationModalProps) => {
     const [temporaryDelivery, setTemporaryDelivery] = useState<Partial<Delivery>>(delivery ?? {});
-    const { partners } = usePartners(direction === DirectionType.Outbound ? 'product' : 'material', temporaryDelivery?.ownMaterialNumber ?? null);
-    const { sites } = useSites();
     const { notify } = useNotifications();
-    const [formError, setFormError] = useState(false);
+    const { openDialog } = useDataModal();
     const dailyDeliveries = useMemo(
         () =>
             deliveries?.filter(
@@ -281,39 +271,7 @@ export const DeliveryInformationModal = ({
         [deliveries, delivery, direction, site]
     );
 
-    const handleSaveClick = () => {
-        temporaryDelivery.customerOrderNumber ||= undefined;
-        temporaryDelivery.customerOrderPositionNumber ||= undefined;
-        temporaryDelivery.supplierOrderNumber ||= undefined;
-        if (!isValidDelivery(temporaryDelivery)) {
-            setFormError(true);
-            return;
-        }
-        setFormError(false);
-
-        postDelivery(temporaryDelivery)
-            .then(() => {
-                onSave();
-                notify({
-                        title: 'Delivery Added',
-                        description: 'The Delivery has been added',
-                        severity: 'success',
-                    },
-                );
-            })
-            .catch((error) => {
-                notify({
-                        title: error.status === 409 ? 'Conflict' : 'Error requesting update',
-                        description: error.status === 409 ? 'Delivery conflicting with an existing one' : error.error,
-                        severity: 'error',
-                    },
-                );
-            })
-            .finally(() => onClose());
-    };
-
     const handleClose = () => {
-        setFormError(false);
         setTemporaryDelivery({});
         onClose();
     };
@@ -332,6 +290,9 @@ export const DeliveryInformationModal = ({
             }
         }
     };
+    const handleEdit = (row: Delivery) => {
+        openDialog('delivery', {...row}, deliveries, 'edit');
+    };
 
     useEffect(() => {
         if (delivery) {
@@ -342,329 +303,37 @@ export const DeliveryInformationModal = ({
         <>
             <Dialog open={open && delivery !== null} onClose={handleClose} data-testid="delivery-modal">
                 <DialogTitle variant="h3" textAlign="center">
-                    {capitalize(mode)} Delivery Information
+                    Delivery Information
                 </DialogTitle>
-                <Stack padding="0 2rem 2rem" sx={{ width: '60rem' }}>
-                    <Grid container spacing={1} padding=".25rem">
-                        {mode === 'create' ? (
-                            <>
-                                <GridItem label="Material Number" value={temporaryDelivery.ownMaterialNumber ?? ''} />
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="ownBpns"
-                                        options={sites ?? []}
-                                        getOptionLabel={(option) => option.name ?? ''}
-                                        isOptionEqualToValue={(option, value) => option?.bpns === value.bpns}
-                                        onChange={(_, value) =>
-                                            setTemporaryDelivery({
-                                                ...temporaryDelivery,
-                                                ...(direction === DirectionType.Outbound
-                                                    ? { originBpns: value?.bpns ?? undefined }
-                                                    : { destinationBpns: value?.bpns ?? undefined }),
-                                            })
-                                        }
-                                        value={
-                                            sites?.find(
-                                                    (s) =>
-                                                        (direction === DirectionType.Outbound
-                                                            ? s.bpns === temporaryDelivery.originBpns
-                                                            : s.bpns === temporaryDelivery.destinationBpns)
-                                                ) ?? null
-                                        }
-                                        label={`${direction === DirectionType.Outbound ? 'Origin' : 'Destination'}*`}
-                                        placeholder={`Select a ${direction === DirectionType.Outbound ? 'Origin' : 'Destination'} Site`}
-                                        error={
-                                            formError &&
-                                            (direction === DirectionType.Outbound ? !temporaryDelivery.originBpns : !temporaryDelivery.destinationBpns)
-                                        }
-                                        data-testid="delivery-own-bpns-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="departure-type"
-                                        options={DEPARTURE_TYPES}
-                                        getOptionLabel={(option) => option.value ?? ''}
-                                        isOptionEqualToValue={(option, value) => option?.key === value.key}
-                                        onChange={(_, value) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, departureType: value?.key ?? undefined })
-                                        }
-                                        value={DEPARTURE_TYPES.find((dt) => dt.key === temporaryDelivery.departureType) ?? null}
-                                        label="Departure Type*"
-                                        placeholder="Select the type of departure"
-                                        error={formError && !temporaryDelivery?.departureType}
-                                        data-testid="delivery-departure-type-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="arrival-type"
-                                        options={ARRIVAL_TYPES}
-                                        getOptionLabel={(option) => option.value ?? ''}
-                                        isOptionEqualToValue={(option, value) => option?.key === value.key}
-                                        onChange={(_, value) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, arrivalType: value?.key ?? undefined })
-                                        }
-                                        value={ARRIVAL_TYPES.find((dt) => dt.key === temporaryDelivery.arrivalType) ?? null}
-                                        label="Arrival Type*"
-                                        placeholder="Select the type of departure"
-                                        error={
-                                            formError &&
-                                            (!temporaryDelivery?.arrivalType ||
-                                                (temporaryDelivery?.arrivalType === 'actual-arrival' &&
-                                                    temporaryDelivery?.departureType !== 'actual-departure'))
-                                        }
-                                        data-testid="delivery-arrival-type-field"
-                                    ></LabelledAutoComplete>
-                                </Grid>
-                                <Grid item xs={6} display="flex" alignItems="end" data-testid="delivery-departure-time-field">
-                                    <DateTime
-                                        label="Departure Time*"
-                                        placeholder="Pick Departure Date"
-                                        locale="de"
-                                        error={
-                                            formError &&
-                                            (!temporaryDelivery.dateOfDeparture ||
-                                                (temporaryDelivery.departureType === 'actual-departure' &&
-                                                    temporaryDelivery.dateOfDeparture > new Date()) ||
-                                                (!!temporaryDelivery.dateOfArrival &&
-                                                    temporaryDelivery.dateOfArrival < temporaryDelivery.dateOfDeparture))
-                                        }
-                                        value={temporaryDelivery?.dateOfDeparture ?? null}
-                                        onValueChange={(date) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, dateOfDeparture: date ?? undefined })
-                                        }
-                                    />
-                                </Grid>
-                                <Grid item xs={6} display="flex" alignItems="end" data-testid="delivery-arrival-time-field">
-                                    <DateTime
-                                        label="Arrival Time*"
-                                        placeholder="Pick Arrival Date"
-                                        locale="de"
-                                        error={
-                                            formError &&
-                                            (!temporaryDelivery?.dateOfArrival ||
-                                                (temporaryDelivery?.arrivalType === 'actual-arrival' &&
-                                                    temporaryDelivery?.dateOfArrival > new Date()))
-                                        }
-                                        value={temporaryDelivery?.dateOfArrival ?? null}
-                                        onValueChange={(date) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, dateOfArrival: date ?? undefined })
-                                        }
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        sx={{ margin: '0' }}
-                                        id="partner"
-                                        options={partners ?? []}
-                                        getOptionLabel={(option) => option?.name ?? ''}
-                                        label="Partner*"
-                                        placeholder="Select a Partner"
-                                        error={formError && !temporaryDelivery?.partnerBpnl}
-                                        onChange={(_, value) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, partnerBpnl: value?.bpnl ?? undefined })
-                                        }
-                                        value={partners?.find((p) => p.bpnl === temporaryDelivery.partnerBpnl) ?? null}
-                                        isOptionEqualToValue={(option, value) => option?.bpnl === value?.bpnl}
-                                        data-testid="delivery-partner-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="partnerBpns"
-                                        options={partners?.find((s) => s.bpnl === temporaryDelivery?.partnerBpnl)?.sites ?? []}
-                                        getOptionLabel={(option) => option.name ?? ''}
-                                        disabled={!temporaryDelivery?.partnerBpnl}
-                                        isOptionEqualToValue={(option, value) => option?.bpns === value.bpns}
-                                        onChange={(_, value) =>
-                                            setTemporaryDelivery({
-                                                ...temporaryDelivery,
-                                                ...(direction === DirectionType.Inbound
-                                                    ? { originBpns: value?.bpns ?? undefined }
-                                                    : { destinationBpns: value?.bpns ?? undefined }),
-                                            })
-                                        }
-                                        value={
-                                            partners
-                                                ?.find((s) => s.bpnl === temporaryDelivery?.partnerBpnl)
-                                                ?.sites.find(
-                                                    (s) =>
-                                                        (direction === DirectionType.Inbound
-                                                            ? s.bpns === temporaryDelivery.originBpns
-                                                            : s.bpns === temporaryDelivery.destinationBpns)
-                                                ) ?? null
-                                        }
-                                        label={`${direction === DirectionType.Inbound ? 'Origin' : 'Destination'}*`}
-                                        placeholder={`Select a ${direction === DirectionType.Inbound ? 'Origin' : 'Destination'} Site`}
-                                        error={
-                                            formError &&
-                                            (direction === DirectionType.Inbound ? !temporaryDelivery.originBpns : !temporaryDelivery.destinationBpns)
-                                        }
-                                        data-testid="delivery-partner-bpns-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Quantity*</FormLabel>
-                                    <Input
-                                        hiddenLabel
-                                        type="number"
-                                        value={temporaryDelivery.quantity ?? ''}
-                                        error={formError && (temporaryDelivery?.quantity == null || temporaryDelivery.quantity <= 0)}
-                                        onChange={(e) =>
-                                            setTemporaryDelivery((curr) => ({
-                                                ...curr,
-                                                quantity: e.target.value ? parseFloat(e.target.value) : undefined,
-                                            }))
-                                        }
-                                        sx={{ marginTop: '.5rem' }}
-                                        data-testid="delivery-quantity-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="uom"
-                                        value={
-                                            temporaryDelivery.measurementUnit
-                                                ? {
-                                                      key: temporaryDelivery.measurementUnit,
-                                                      value: getUnitOfMeasurement(temporaryDelivery.measurementUnit),
-                                                  }
-                                                : null
-                                        }
-                                        options={UNITS_OF_MEASUREMENT}
-                                        getOptionLabel={(option) => option?.value ?? ''}
-                                        label="UOM*"
-                                        placeholder="Select unit"
-                                        error={formError && !temporaryDelivery?.measurementUnit}
-                                        onChange={(_, value) => setTemporaryDelivery((curr) => ({ ...curr, measurementUnit: value?.key }))}
-                                        isOptionEqualToValue={(option, value) => option?.key === value?.key}
-                                        data-testid="delivery-uom-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Tracking Number</FormLabel>
-                                    <Input
-                                        id="tracking-number"
-                                        hiddenLabel
-                                        type="text"
-                                        value={temporaryDelivery?.trackingNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, trackingNumber: event.target.value })
-                                        }
-                                        sx={{ marginTop: '.5rem' }}
-                                        data-testid="delivery-tracking-number-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <LabelledAutoComplete
-                                        id="incoterm"
-                                        value={
-                                            temporaryDelivery.incoterm
-                                                ? INCOTERMS.find((i) => i.key === temporaryDelivery.incoterm) ?? null
-                                                : null
-                                        }
-                                        options={INCOTERMS.filter((i) =>
-                                            direction === DirectionType.Inbound ? i.responsibility !== 'supplier' : i.responsibility !== 'customer'
-                                        )}
-                                        getOptionLabel={(option) => option?.value ?? ''}
-                                        label="Incoterm*"
-                                        placeholder="Select Incoterm"
-                                        error={formError && !temporaryDelivery?.incoterm}
-                                        onChange={(_, value) => setTemporaryDelivery((curr) => ({ ...curr, incoterm: value?.key }))}
-                                        isOptionEqualToValue={(option, value) => option?.key === value?.key}
-                                        data-testid="delivery-incoterm-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Customer Order Number</FormLabel>
-                                    <Input
-                                        id="customer-order-number"
-                                        type="text"
-                                        error={formError && !isValidOrderReference(temporaryDelivery)}
-                                        value={temporaryDelivery?.customerOrderNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, customerOrderNumber: event.target.value })
-                                        }
-                                        data-testid="delivery-customer-order-number-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Customer Order Position</FormLabel>
-                                    <Input
-                                        id="customer-order-position-number"
-                                        type="text"
-                                        error={formError && !isValidOrderReference(temporaryDelivery)}
-                                        value={temporaryDelivery?.customerOrderPositionNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryDelivery({
-                                                ...temporaryDelivery,
-                                                customerOrderPositionNumber: event.target.value,
-                                            })
-                                        }
-                                        data-testid="delivery-customer-order-position-field"
-                                    />
-                                </Grid>
-                                <Grid item xs={6}>
-                                    <FormLabel>Supplier Order Number</FormLabel>
-                                    <Input
-                                        id="supplier-order-number"
-                                        type="text"
-                                        error={
-                                            formError && !!temporaryDelivery.supplierOrderNumber && !temporaryDelivery.customerOrderNumber
-                                        }
-                                        value={temporaryDelivery?.supplierOrderNumber ?? ''}
-                                        onChange={(event) =>
-                                            setTemporaryDelivery({ ...temporaryDelivery, supplierOrderNumber: event.target.value })
-                                        }
-                                        data-testid="delivery-supplier-order-number-field"
-                                    />
-                                </Grid>
-                            </>
-                        ) : (
-                            <Grid item xs={12}>
-                                {
-                                    <Table
-                                        title={`${direction === DirectionType.Outbound ? 'Outgoing Shipments' : 'Incoming Deliveries'} on ${
-                                            direction === DirectionType.Outbound
-                                                ? new Date(temporaryDelivery.dateOfDeparture!).toLocaleDateString('en-UK', {
-                                                      weekday: 'long',
-                                                      day: '2-digit',
-                                                      month: '2-digit',
-                                                      year: 'numeric',
-                                                  })
-                                                : new Date(temporaryDelivery.dateOfArrival!).toLocaleDateString('en-UK', {
-                                                    weekday: 'long',
-                                                    day: '2-digit',
-                                                    month: '2-digit',
-                                                    year: 'numeric',
-                                                })
-                                        }`}
-                                        density="standard"
-                                        getRowId={(row) => row.uuid}
-                                        columns={createDeliveryColumns(handleDelete)}
-                                        rows={dailyDeliveries}
-                                        hideFooter
-                                    />
-                                }
-                            </Grid>
-                        )}
+                <Stack padding="0 2rem 2rem" sx={{ maxWidth: '95vw' }}>
+                    <Grid item xs={12}>
+                        <Table
+                            title={`${direction === DirectionType.Outbound ? 'Outgoing Shipments' : 'Incoming Deliveries'} on ${
+                                direction === DirectionType.Outbound
+                                    ? new Date(temporaryDelivery.dateOfDeparture!).toLocaleDateString('en-UK', {
+                                            weekday: 'long',
+                                            day: '2-digit',
+                                            month: '2-digit',
+                                            year: 'numeric',
+                                        })
+                                    : new Date(temporaryDelivery.dateOfArrival!).toLocaleDateString('en-UK', {
+                                        weekday: 'long',
+                                        day: '2-digit',
+                                        month: '2-digit',
+                                        year: 'numeric',
+                                    })
+                            }`}
+                            density="standard"
+                            getRowId={(row) => row.uuid}
+                            columns={createDeliveryColumns(handleDelete, handleEdit)}
+                            rows={dailyDeliveries}
+                            hideFooter
+                        />
                     </Grid>
                     <Box display="flex" gap="1rem" width="100%" justifyContent="end" marginTop="2rem">
                         <Button variant="outlined" color="primary" sx={{ display: 'flex', gap: '.25rem' }} onClick={handleClose}>
                             <Close></Close> Close
                         </Button>
-                        {mode === 'create' && (
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                sx={{ display: 'flex', gap: '.25rem' }}
-                                onClick={() => handleSaveClick()}
-                                data-testid="save-delivery-button"
-                            >
-                                <Save></Save> Save
-                            </Button>
-                        )}
                     </Box>
                 </Stack>
             </Dialog>
