+++ frontend/src/features/material-details/components/MaterialDetails.tsx
@@ -24,9 +24,9 @@ import { Box, capitalize, Stack, Typography } from '@mui/material';
 import { MaterialDetailsHeader } from './MaterialDetailsHeader';
 import { OwnSummaryPanel } from './SummaryPanel';
 import { CollapsibleSummary } from './CollapsibleSummary';
-import { DataCategory, useMaterialDetails } from '../hooks/useMaterialDetails';
+import { useMaterialDetails } from '../hooks/useMaterialDetails';
 import { useNotifications } from '@contexts/notificationContext';
-import { useDataModal } from '@contexts/dataModalContext';
+import { DataCategory, useDataModal } from '@contexts/dataModalContext';
 import { ReactNode, useCallback, useEffect, useMemo, useState } from 'react';
 import { groupBy } from '@util/helpers';
 import { DirectionType } from '@models/types/erp/directionType';
