+++ frontend/src/services/delivery-service.ts
@@ -38,6 +38,21 @@ export const postDelivery = async (delivery: Partial<Delivery>) => {
   }
   return res.json();
 }
+export const updateDelivery = async (delivery: Partial<Delivery>) => {
+  const res = await fetch(config.app.BACKEND_BASE_URL + config.app.ENDPOINT_DELIVERY, {
+    method: 'PUT',
+    body: JSON.stringify(delivery),
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': `Bearer ${AuthenticationService.getToken()}`
+    },
+  });
+  if(res.status >= 400) {
+    const error = await res.json();
+    throw error;
+  }
+  return res.json();
+}
 
 export const deleteDelivery = async (id: UUID) => {
   const res = await fetch(config.app.BACKEND_BASE_URL + config.app.ENDPOINT_DELIVERY + `/${id}`, {
